Create README.md
To run our program, you must compile and run each of the classes which have a main method in it but the client class should be compiled last. Enter 1 to connect to hiscinema.com authoritative server or enter 2 to herCDN.com authoritative server. 

Then, you can enter a link (e.g. video.netcinema.com) and the value for it in the record in the hiscinema authoritative server is going to be displayed.